+++ buildSrc/src/main/kotlin/com/danrusu/pods4k/immutableArrays/core/ImmutableArrayGenerator.kt
@@ -3,7 +3,6 @@ package com.danrusu.pods4k.immutableArrays.core
 import com.danrusu.pods4k.immutableArrays.BaseType
 import com.danrusu.pods4k.immutableArrays.BaseType.GENERIC
 import com.danrusu.pods4k.immutableArrays.ImmutableArrayConfig
-import com.danrusu.pods4k.immutableArrays.createImmutableArrayBuilder
 import com.danrusu.pods4k.utils.ParameterDSL
 import com.danrusu.pods4k.utils.addGenericTypes
 import com.danrusu.pods4k.utils.addPrimaryConstructor
@@ -761,25 +760,28 @@ private fun TypeSpec.Builder.addFilter(baseType: BaseType) {
         kdoc = "Returns an immutable array containing only the elements matching the given [predicate].",
         modifiers = listOf(KModifier.INLINE),
         name = "filter",
-        parameters = { "predicate"(type = lambda<Boolean> { "element"(type = baseType.type) }) },
+        parameters = {
+            "predicate"(
+                modifiers = listOf(KModifier.CROSSINLINE),
+                type = lambda<Boolean> { "element"(type = baseType.type) },
+            )
+        },
         returns = baseType.getGeneratedTypeName(),
     ) {
-        createImmutableArrayBuilder(name = "result", forType = baseType)
-        controlFlow("for (element in values)") {
-            controlFlow("if (predicate(element))") {
-                statement("result.add(element)")
-            }
-        }
-        statement("if (result.size == size) return this")
-        emptyLine()
-        statement("return result.build()")
+        comment("delegate to filterIndexed as that's extremely optimized")
+        statement("return filterIndexed { _, element -> predicate(element) }")
     }
 }
 
 private fun TypeSpec.Builder.addFilterIndexed(baseType: BaseType) {
     function(
-        kdoc = "Returns an immutable array containing only the elements matching the given [predicate].",
-        modifiers = listOf(KModifier.INLINE),
+        kdoc = """
+            Returns an immutable array containing only the elements matching the given [predicate].
+
+            Warning: This code is quite dense because it's highly optimized to reduce the temporary
+            memory overhead and also dramatically improves performance due to branch elimination and
+            a perfectly-sized result.
+        """.trimIndent(),
         name = "filterIndexed",
         parameters = {
             "predicate"(
@@ -791,15 +793,67 @@ private fun TypeSpec.Builder.addFilterIndexed(baseType: BaseType) {
         },
         returns = baseType.getGeneratedTypeName(),
     ) {
-        createImmutableArrayBuilder(name = "result", forType = baseType)
-        controlFlow("forEachIndexed { index, element ->") {
-            controlFlow("if (predicate(index, element))") {
-                statement("result.add(element)")
-            }
-        }
-        statement("if (result.size == size) return this")
+        statement("if (isEmpty()) return this")
         emptyLine()
-        statement("return result.build()")
+        comment("divide by 32 rounding up")
+        statement("val bitArraySize = (size + 31) ushr 5")
+
+        // Note that I tried using a Long array but that was marginally slower and more complex due to int conversions
+        comment("store an array of int values whose 1-bits capture which elements pass the predicate")
+        statement("val bitArray = IntArray(bitArraySize)")
+        emptyLine()
+        statement("var resultSize = 0")
+        comment("the bit index into the current 32-bit int")
+        statement("var bitIndex = -1 // start at -1 as it gets incremented right away in the loop")
+        statement("var bitArrayIndex = 0")
+        statement("var currentBits = 0 // the current 32-bits with no elements yet")
+        controlFlow("forEachIndexed { index, element ->") {
+            controlFlow("if (++bitIndex == 32)") {
+                comment("reached the end of the current bits so store them and reset")
+                statement("bitArray[bitArrayIndex++] = currentBits")
+                statement("currentBits = 0")
+                statement("bitIndex = 0")
+            }
+            comment("jit turns this into a branchless operation")
+            statement("val currentElement = if (predicate(index, element)) 1 else 0")
+            emptyLine()
+            comment("conditionally increase the size without branching")
+            statement("resultSize += currentElement")
+            emptyLine()
+            comment("conditionally include the current element without branching")
+            statement("currentBits = currentBits or (currentElement shl bitIndex)")
+        }
+        statement("if (resultSize == 0) return EMPTY")
+        statement("if (resultSize == size) return this")
+        emptyLine()
+        comment("store the last set of partially-filled bits")
+        statement("bitArray[bitArrayIndex] = currentBits")
+        emptyLine()
+        if (baseType == GENERIC) {
+            suppress("UNCHECKED_CAST")
+            statement("val result = arrayOfNulls<Any>(resultSize) as Array<T>")
+        } else {
+            statement("val result = ${baseType.backingArrayConstructor}(resultSize)")
+        }
+        statement("var resultIndex = 0")
+        statement("bitArrayIndex = 0")
+        statement("bitIndex = -1")
+        statement("currentBits = bitArray[0]")
+        statement("var originalIndex = 0")
+        comment("check the resultIndex instead of the originalIndex so that we can stop early")
+        controlFlow("while (resultIndex < result.size)") {
+            controlFlow("if (++bitIndex == 32)") {
+                comment("reached the end of the current bits so get the next 32 bits and reset")
+                statement("currentBits = bitArray[++bitArrayIndex]")
+                statement("bitIndex = 0")
+            }
+            comment("always copy to avoid branching as resultIndex won't increment if current element isn't included")
+            statement("result[resultIndex] = this[originalIndex++]")
+            comment("increment the resultIndex if the current element should be included")
+            statement("val currentElement = (currentBits ushr bitIndex) and 1")
+            statement("resultIndex += currentElement")
+        }
+        statement("return ${baseType.generatedClassName}(result)")
     }
 }
 
@@ -808,18 +862,16 @@ private fun TypeSpec.Builder.addFilterNot(baseType: BaseType) {
         kdoc = "Returns an immutable array containing only the elements that don't match the [predicate].",
         modifiers = listOf(KModifier.INLINE),
         name = "filterNot",
-        parameters = { "predicate"(type = lambda<Boolean> { "element"(type = baseType.type) }) },
+        parameters = {
+            "predicate"(
+                modifiers = listOf(KModifier.CROSSINLINE),
+                type = lambda<Boolean> { "element"(type = baseType.type) },
+            )
+        },
         returns = baseType.getGeneratedTypeName(),
     ) {
-        createImmutableArrayBuilder(name = "result", forType = baseType)
-        controlFlow("for (element in values)") {
-            controlFlow("if (!predicate(element))") {
-                statement("result.add(element)")
-            }
-        }
-        statement("if (result.size == size) return this")
-        emptyLine()
-        statement("return result.build()")
+        comment("delegate to filterIndexed as that's extremely optimized")
+        statement("return filterIndexed { _, element -> !predicate(element) }")
     }
 }
 
@@ -1038,6 +1090,7 @@ private fun TypeSpec.Builder.addDistinctBy(baseType: BaseType) {
         name = "distinctBy",
         parameters = {
             "selector"(
+                modifiers = listOf(KModifier.CROSSINLINE),
                 type = lambda(
                     parameters = { "element"(type = baseType.type) },
                     returnType = key,